Let scripts set scores by using float values BUG: 139311
svn path=/trunk/extragear/multimedia/amarok/; revision=617391

+++ src/amarokcore/amarokdcophandler.cpp
@@ -430,7 +430,7 @@ namespace Amarok
         CollectionDB::instance()->setSongPercentage(url, score);
     }
 
-    void DcopPlayerHandler::setScoreByPath( const QString &url, int score )
+    void DcopPlayerHandler::setScoreByPath( const QString &url, float score )
     {
         CollectionDB::instance()->setSongPercentage(url, score);
     }
